第十七次提交：使用拦截器结合Threadlocal完成token处理 拦截器鉴权合法对请求放行 1.解析token,获取id和手机号码，构造user对象，存入ThreadLocal Claims claims = JwtUtils.getClaims(token); String mobile = (String) claims.get("phone"); Integer id = (Integer) claims.get("id"); 存入ThreadLocal(UserHolder.set(user);)
编写Threadloca工具类，提供线程存储和获取对象的方法
拦截器解析token验证通过后，将对象存入Threadlocal中
修改控制器方法，从ThreadLocal获取当前登录用户对象

+++ tanhua-app-server/src/main/java/com/tanhua/server/service/UserInfoService.java
@@ -32,7 +32,7 @@ public class UserInfoService {
         userInfoApi.save(userInfo);
     }
 
-    public void updateHead(MultipartFile headPhoto, Integer id) {
+    public void updateHead(MultipartFile headPhoto, Long id) {
 
         try {
             //1.根据headPhoto上传图片到阿里云,并返回图片url地址
@@ -46,7 +46,8 @@ public class UserInfoService {
                 //2.2包含人脸，调用Api更新
                 UserInfo userInfo = new UserInfo();
                 //设置id
-                userInfo.setId(Long.valueOf(id));
+                //userInfo.setId(Long.valueOf(id));
+                userInfo.setId(id);
                 //设置头像链接地址
                 imageUrl="https://tanhua001.oss-cn-beijing.aliyuncs.com/2021/04/19/a3824a45-70e3-4655-8106-a1e1be009a5e.jpg";
                 userInfo.setAvatar(imageUrl);
